Refactoring: Input Label.
